Add Prometheus 1.x
Dependencies were upgraded to the 1.x versions of the Prometheus client
artifacts in micrometer-registry-prometheus.
This is a breaking change compared to the code that was there before.
This is unavoidable due to the breaking changes
in the Prometheus client from 0.x to 1.x.
The previous code from micrometer-registry-prometheus was moved to
a new module micrometer-registry-prometheus-simpleclient that
is deprecated but available to give users a backward compatible option
in case they need to make changes to use the new Prometheus client.

The base package used in micrometer-registry-prometheus was changed to
io.micrometer.prometheusmetrics to differentiate it from the
prior package now used in micrometer-registry-prometheus-simpleclient.
This avoids split packages and allows both modules to be used
in the same application even.

Tests were adapted to the new API and differences in the scrape output.
Tests involving exemplars are commented out until
exemplar support is added.

PrometheusHistogram's cumulative buckets were converted to delta in the
registry. The new Prometheus client wants delta buckets which is weird
since it converts them back to cumulative internally.

PrometheusNamingConvention behavior was changed, it does not append
"_total" to counters anymore. The new Prometheus client validates
Counter names and check if they end with "_total".
If they don't, it appends "_total". If they do, it throws an exception:
java.lang.IllegalArgumentException: 'cnt_total': Illegal metric name...
Because of this behavior, if we want to use PrometheusNamingConvention,
we need to modify it so that it does not append "_total".

Known issues in the micrometer-registry-prometheus module after this PR:
- Exemplars and Native Histograms do not work.
- VictoriaMetrics histograms are not supported.
- The OSGi test was failing when the new Prometheus client was used, so
it was updated to use the -simpleclient module for now.

See gh-3987
See gh-4406

Co-authored-by: Tommy Ludwig <8924140+shakuzen@users.noreply.github.com>