Sanitize metric names for the Prometheus client
We need to sanitize metric names for the Prometheus client otherwise
it fails with something like this:
java.lang.IllegalArgumentException: 'jvm_info': Illegal metric name.
The metric name must not include the '_info' suffix.

This can be reproduced with:
Gauge.builder("test.info", () -> 1).register(registry);

The name of this Gauge will be test_info in Prometheus that is invalid
according to the new client, see PrometheusNaming:
https://github.com/prometheus/client_java/blob/8bb75446c6dca2a4e2c074e39fe34c1b061b028a/prometheus-metrics-model/src/main/java/io/prometheus/metrics/model/snapshots/PrometheusNaming.java#L37-L40

We have built-in instrumentation that fails this validation too, see JvmInfoMetrics:
https://github.com/micrometer-metrics/micrometer/blob/966f5fe953a69c315b44bf4f63adf101a060bbfb/micrometer-core/src/main/java/io/micrometer/core/instrument/binder/jvm/JvmInfoMetrics.java#L32

If we don't want to break this, we need to sanitize the invalid suffixes.

Please check the tests because the behavior can be unexpected sometimes:
- If you create a Micrometer Gauge named test.info,
the name of the Prometheus gauge will be test_info (since the registry
removes but the Prometheus client adds back the `_info` suffix).
- If you create a Micrometer Counter named test.total,
the name of the Prometheus counter will be test_total (since the
registry removes but the Prometheus client adds back the _total suffix).
- If you create a Micrometer Counter named test.info,
the name of the Prometheus counter will be test_total (since the
registry removes the _info suffix and the Prometheus client adds the
_total suffix for counters).
- If you create a Micrometer Gauge named test.total,
the name of the Prometheus counter will be test (since the registry
removes the _total suffix and the Prometheus client does not add
_info suffix for gauges).
- Similarly, _created and _bucket will be removed otherwise
the the Prometheus client will fail.