Unmark stale meters that are unchanged upon register
Registering a MeterFilter will mark all existing meters as stale, which causes filters to be applied again next time the meter is registered. But the meter may not be changed by the newly registered filter. In this case, it was not being unmarked as stale, which prevented a shortcut for performance.

See https://github.com/micrometer-metrics/micrometer/pull/4857#discussion_r1555861042