Fix AspectJ load-time weaving and class-level annotations
The pointcut syntax used in the aspects is not valid for the AspectJ
weaver/compiler. Replaced `and not` with `&& !` to fix it and also had
to add a condition to match only methods and not constructors which the
implementation assumed it gets.

One problem was that the aspects were only tested with Spring AOP
proxies, but AspectJ weaver/compiler bugs were going unnoticed.  The
added LTW test module will run with the `aspectjweaver` java agent to
weave the classes at load time, then tests if the aspects are applied
and metrics collected as expected.