Fix histogram consistency in PrometheusMeterRegistry
The Histogram implementation of PrometheusMeterRegistry lets
inconsistencies to happen in certain concurrent scenarios. For example,
if a Timer is scraped while a data point is recorded,
the count field of the Timer can be less than the value of the
last bucket in its histogram snapshot
(the bucket is already incremented while the counter is not).

With the new Prometheus Client (1.x), this can cause exceptions since:
1. The Prometheus Client's ClassicHistogramBuckets is not cumulative
but the histograms we use are so we need to convert it to
non-cumulative (delta) by doing a little math.
2. If missing, we calculate the value of the +Inf bucket using
the count and the last bucket of the histogram, the calculation looks
like this: +Inf bucket = count - last bucket (delta).
3. In this situation, if the value of the count is less than the value
of the last bucket in the histogram snapshot (concurrency issue),
the value of the +Inf bucket will be negative which is validated
by the Prometheus Client and an exception is thrown:
java.lang.IllegalArgumentException: Counts in ClassicHistogramBuckets cannot be negative.

In case of Prometheus Client 0.x, there is no such exception since
we don't need to calculate the value of the +Inf bucket but we reuse
the cumulative value of count.

Closes gh-5193