Do not register GC metrics when GC notifications are unavailable (#5149)
This addresses the issue that GraalVM currently does not support GC notifications which means no useful GC metrics with Micrometer. Micrometer does not provide any hint and even registers a few useless metrics which only have initial values but never get updates as this relies on GC notifications.

Instead, a log letting users know GC notifications are unavailable is emitted and this now skips registering the GC-related metrics that can't be updated.