Make PrometheusMeterRegistryTest less flaky
Tests that also include verification for Exemplars
in PrometheusMeterRegistryTest can be flaky since the Exemplars sampler
has a rate limiter that does not let sampling happening too frequently.
Since our tests want to verify (sampled) exemplars, we set the
rate-limit interval to it's minimum
(io.prometheus.exemplars.sampleIntervalMilliseconds=1(ms)) and sleep
a little between recordings to not hit the rate-limit. It seems the
amount of sleep is not always enough and we need to sleep more.

Closes gh-5302