Flush channel on close so buffered metrics are not lost (#5195)
We thought disposing the connection was enough to flush any buffered metrics, but this was not flushing the channel. This was missed by our IT tests (StatsdMeterRegistryPublishTest) because we did not use buffered config. This makes sure to flush the channel before disposing the connection so any buffered statsd lines are sent on close.

Resolves gh-2141