Add history-tracking to ObservationValidator
This change lets you peek into the history of certain interactions with
your Observations if an InvalidObservationException is thrown.
This includes which method was called on your Observations
(start, stop, error, etc) and also the relevant part of the stack trace.

InvalidObservationException can provide you the full history through its
getHistory method and it also gives you a summary through its toString.