Remove registered meters when shutting down TimedExecutorService (#5367)
Applications may have use cases where they wish to recreate an ExecutorService with different properties and monitor it using the same name as the previous instance. To allow for this, track registered meters and remove them when shutting down the wrapped ExecutorService (TimedExecutorService). Note that this will not remove the registered meters in cases where the ExecutorService is not wrapped or the wrapped service is not the one that is shutdown.

Resolves gh-5366