Virtual threads metrics (#5067)
Creates a new module micrometer-java21 for code that needs to be baselined on Java 21, including the added virtual threads metrics. The instrumentation is JFR-based, although that is an implementation detail. There are some configuration options for the instrumentation but they are not exposed as public API until there is a clear need for that from users.

Fixes gh-3956

Co-authored-by: Jonatan Ivanov <jonatan.ivanov@gmail.com>