Run concurrency tests as part of CI build (#5498)
While we added concurrency tests, they were not being run as part of the build. In order to ensure concurrency issues aren't introduced by a change we didn't expect to affect that, we should regularly run the concurrency test suite.

The `default` mode took too long on CI, but `quick` seems to be reasonably quick. If a concurrency issue is harder to reproduce, running the tests with a more stressed mode may be necessary.