Consistent tagging and propagate sendAsync throwable (#5553)
When handling the sendAsync call in our instrumentation, we were not propagating the throwable observed to the CompletionFuture returned. This fixes that and adds a test. To keep tagging consistent in the case where response is null or not, this also updates tagging to always include `status` and `outcome`. In case the response is null, they will have value `UNKNOWN` for now. This is a change from before for this case, but the behavior before is problematic for the Prometheus registry that expects a consistent set of tags. The change is adding tags that weren't there before in the case response is null, so it should generally not break users.