Add benchmarks for DefaultLTT start/stop (#5595)
* Add benchmarks for DefaultLTT start/stop

Start and stop are called on the critical path. We should have benchmarks for them to evaluate changes that may affect their performance.

* Stop a random sample instead of middle sample

We should get a better idea of average performance with a random sample rather than a sample in a fixed position in the active tasks collection.

* AverageTime BenchmarkMode instead of single-shot

Using the invocation-level setup method, we can use average-time rather than single shot.