Fix Stackdriver Distribution count and bucket count sum mismatch
The count passed to the `HistogramSnapshot` is from the StepTimer and is step-based, while the histogram buckets have a separate time window. This mismatch in time windows caused the two to generally not match which causes a validation error when publishing Distribution metrics to Stackdriver.

This solves this by tracking an infinity bucket in the histogram so we can get a consistent count in the same time window as the rest of the histogram.