Test HTTP client instrumentation handles path template on 404 response
Instrumentation was recently updated to not replace the path with a constant such as NOT_FOUND on 404 response. This tests that instrumentation adheres to that. The deprecated Jetty client instrumentation based on the older version of Jetty client does not adhere, but since it is deprecated, the test was disabled rather than updating the deprecated instrumentation.

Closes gh-5812