Handle programatically added log4j2 LoggerConfig (#5902)
The goal of this is to be able to bind to a newly added `LoggerConfig` even if the `Configuration` is the same. This means we always need to check if the filter was already added to avoid adding it twice. We need to keep track of if anything was changed to avoid an endless loop when we call `updateLoggers()`

Signed-off-by: Patrik Ivarsson <patrik.ivarsson@avanza.se>