Send 0 mean when count is 0
The Stackdriver backend rejects Distributions where the mean is non-zero when count is 0. The mean is calculated from a different time window than the histogram, which is what makes this situation possible in the first place. Changing that is too big of scope for fixing this bug. Rather, this checks for the count being 0 and sets the mean to 0 in that case.