Revert "Bump org.aspectj:aspectjweaver from 1.9.22.1 to 1.9.23 (#6020)"
This reverts commit 529b5ea993722eabe583029e35e8589f7c49a450.

AspectJ does not use semantic versioning so the third number
in the version is not the patch version but it indicates the latest
Java version that is supported (the 4th number is the patch version).
Other than we should not do this in our patch releases,
AspectJ 1.9.23 needed to upgrade the Eclipse JDT compiler in order to
support Java 23 which drops compile-time weaving support for Java 8.

See gh-6020