Delete deprecated JDK 11 classes
These were deprecated and moved to the micrometer-java11 module with Micrometer 1.13.0. The multi-release JAR arrangement that is used for these in micrometer-core has some issues, which is what caused the move to micrometer-java11. See gh-4728. It would help simplify our build to remove the multi-release JAR plugin, which ends up downloading a JDK for the target release on CI builds even if the configured toolchain can target that release.