Replace global dependency management with version catalog
Completely removes dependencies.gradle which acted as a global dependency management file for the whole build. We previously introduced a version catalog and these changes make full use of it. Individual dependency versions are managed in the catalog and the dependencies are referenced using the type-safe API where they are declared. This should alleviate confusion contributors have about where to put versions and how to declare dependencies. It also avoids needless dependency management in all sub-projects even for dependencies that are only used in one place. This also removes the global dependency management via the BOMs that were declared in dependencies.gradle. Instead, the BOMs are declared in the projects that use them at the appropriate scope. An exception to this is the JUnit BOM because each 'java-library' project should have tests and use JUnit, so it is added to all of them from the root build.gradle.