Fix removing dependencyManagement from POMs
There was always a dependencyManagement section due to the previous setup where dependencies.gradle applied dependency management to each project. Now that that is gone, some projects do not have a dependencyManagement section when the POMs are processed. The logic before only checked for null but the object was an empty collection in the case the dependencyManagement node wasn't found. This change correctly checks the dependencyManagement node(s) object is not null and not empty.

See gh-1819
See gh-6080