Remove Spring Boot 2 samples
Spring Boot 2 is out of OSS support at this point, and the latest versions of Micrometer are used with Spring Boot 3, so it is odd to have samples with the latest version of Micrometer using Spring Boot 2. Therefore, I don't think they add much value and can be removed. There is no code run as part of the build (only compilation), and we should try to eliminate dependency cycles: these samples depend on Spring but Spring depends on Micrometer. If we still want Spring-based samples, a separate repository would be a better setup.