Remove Hazelcast 3 tests (#6150)
This simplifies our version catalog so we don't need to track upgrades to two different versions of the same artifact, with which Dependabot notoriously struggles. Hazelcast 3 hasn't had a release in almost 3 years at this point, so I think we don't need to keep testing compatibility with it. Hazelcast 5 is the latest, which is what we compile against.