Track mapped Meter.Id in MultiGauge
MultiGauge maintains the set of ids of the Gauges it created. These ids
are "pre-filtered" which means it stores the ids before the possible
transformations done by MeterFilters.
This changes the behavior and makes MultiGauge tracking the ids after
the MeterFilter transformations. This also means that after this change
MeterFilters can run twice.