Fix ConcurrentModificationException in Exponential Histogram (#6363)
Adds a concurrency test that reproduces the reported case (and another that wasn't reported). Fixes by making the respective operations thread safe. This should have minimal impact on performance because these operations are not frequently done while recording metrics. IndexProviderFactory.getIndexProviderForScale is only called during recording to the histogram when down-scaling. DefaultExponentialHistogramSnapShot.getEmptySnapshotForScale is potentially called when getting a snapshot of the histogram, which will happen when publishing.

Signed-off-by: Lenin Jaganathan <lenin.jaganathan@gmail.com>