Always configure histogram for Stackdriver (#6452)
A distribution is always published to Stackdriver for Timer and DistributionSummary meters. We want this Distribution to have an accurate count in the time window of the Distribution (histogram). A no-op Histogram was being used when no buckets and no percentiles were configured, and that always has a zero count. We can't use the step count because that would not be the count for the same period of time. This will add some overhead to Timer and DistributionSummary meters that did not have any buckets or percentiles configured.

Fixes gh-6401