Remove Dropwizard 5 usage (#6468)
It seems the 5.0.0 release was a mistake and we should not have been using it all this time. The 5.0.1 is the first intended Dropwizard 5 release, and it bumps the Java baseline. We can consider how to deal with that on `main`, but it's best to remove it from maintenance branches in the meantime.