Flag to register JCache cache.removals as FunctionCounter
Some other CacheMeterBinder implementations are known to register the `cache.removals` meter as a FunctionCounter, which is more semantically accurate. For backward compatibility, we do not want to unconditionally change the meter type for everyone, especially in a maintenance release. However, we want to make it possible to register the `cache.removals` as a `FunctionCounter` to align it. In the future, we can flip the default for this flag.

See gh-6434