Register JCache cache.removals as a FunctionCounter by default
This switches the default of the `registerCacheRemovalsAsFunctionCounter` flag from `false` to `true`. This may be a breaking change depending on the backend used and the usage of this metric in queries. This change makes the behavior what we want by default while still allowing users to fall back to the previous behavior by setting the flag to `false` to maintain compatibility with the behavior in prior versions.

See gh-6471
Resolves gh-2754