Add getAll to Getter for repeated metadata (#6312)
It's known that some metadata can be repeated on a carrier. W3C baggage is one example of this in HTTP request headers. The current get API returns a String, and for performance reasons we may not want to check all headers to combine repeated headers when we don't expect there should be repeated headers.

This updates the one non-deprecated main code implementation of Getter (in ObservationGrpcServerInterceptor) to implement the new getAll method.