Provide meter count in HighCardinalityTagsDetector (#6510)
Enhance HighCardinalityTagsDetector to include info about the meter count instead of only giving the meter name. This will be more flexible to adapt in the future with a dedicated type returned now (HighCardinalityMeterInfo). There were some issues with a more simple implementation due to backward compatibility and type erasure and lambdas with the `Consumer` received and `Optional` returned from public APIs.

Signed-off-by: Johnny Lim <izeye@naver.com>
Co-authored-by: Tommy Ludwig <8924140+shakuzen@users.noreply.github.com>