Enhance logging of negative amount in AbstractTimer.record() by printing stack trace (#6530)
In order to help identify the code that is resulting in calling record with a negative amount, include the stacktrace in the log.
Nothing will be logged after the first time if debug level logging is not enabled. We can avoid the cost of instantiating the exception in that case.

---------

Signed-off-by: thonra <dc_soong@outlook.com>
Co-authored-by: Tommy Ludwig <8924140+shakuzen@users.noreply.github.com>