Apache HC 5 OTel semantic convention (#6591)
* Apache HC 5 OTel semantic convention

Adds an ObservationConvention implementation for the Apache HC 5 instrumentation that follows the latest (at this time) OpenTelemetry semantic conventions for HTTP clients.
Also adds an ObservationDocumentation for it as well.

* Make HttpMethods utility class

This will be needed by other HTTP instrumentation.

* Fix copyright year

* Ensure int type attributes have integer values

Specifically, status and port are defined as ints. Uses 0 and -1 for the case when they are unknown/undefined.

* Add documentation and test demonstrating usage