Add MeterFilter.forMeters utility method (#6594)
* Add MeterFilter.forPrefix(..) utility method

I have noticed that I often want to configure MeterFilters only for a selection of meters that start with a given prefix. For example, I might want to ignore tags for a selection of metrics, but not for all of them. The "normal" approach would be to write a custom MeterFilter that checks the name before filtering. The new utility method helps with this task.

Signed-off-by: Peter Jeschke <peter@jeschke.dev>

* Add MeterFilter.forMeters utility method

---------

Signed-off-by: Peter Jeschke <peter@jeschke.dev>
Co-authored-by: Jonatan Ivanov <jonatan.ivanov@gmail.com>