Skip DistributionStatisticsConfig creation when retrieving timers
In the case where the Timer/LongTaskTimer builder is not used, the DistributionStatisticsConfig is necessarily one with the default configuration being passed to registerMeterIfNecessary. We can avoid creating and configuring a DistributionStatisticsConfig that will always have the same config by keeping a static default one to use in these cases.

In performance critical areas the API to register Timer/LongTaskTimer that doesn't use the Builder should be preferred for this reason.

The usage in DefaultMeterObservationHandler has been updated accordingly, which should improve the performance and allocations for all applications using it.