Avoid capturing lambda allocation when retrieving existing meters (#6670)
In particular focuses on Timer, LongTaskTimer, and Counter as these are the types expected to be retrieved (non-functional) the most on hot code paths. Avoiding this allocation makes the retrieval faster and generates less garbage, causing less GC pressure.

Adds to MeterRegistry new public overloads of timer, longTaskTimer, and counter methods that take `Tags` because the existing methods take `Iterable<Tag>` which means `Tags.of` needs to be called on it, which can result in allocation (to check the size of the Iterable) before JIT optimization.

Adds tests to verify the expected allocation behavior of these methods to prevent regressions. It can be particularly difficult to notice allocation caused by capturing lambdas with code review alone.