Improve no-op behavior of the Observation API (#6700)
There are four cases when an Observation can be no-op:
1. The ObservationRegistry is null
2. The ObservationRegistry is ObservationRegistry.NOOP
3. The ObservationRegistry does not have any handlers
4. An Observation is disabled by an ObservationPredicate

In the first three cases, no Observations are created.
In the last case, some Observations can be created but
some can be ignored. In this scenario, we need to support
scoping (context-propagation) but in the first three we don't.

This change makes the NoopObservation truly no-op to cover the
first three cases and introduces NoopButScopeHandlingObservation
to cover the last one.

This should result in better performance in cases where JIT
can eliminate the code because of NoopObservation being truly no-op.