Introduce meter conventions and apply to some JVM metrics (#6682)
Enable specifying a convention to control well-known metrics' names and tags.
Instrumentation can use this so its users can swap the convention applied without having to rewrite the instrumentation or resort to `MeterFilter` which can be more brittle.
The first use of this is in some of the JVM metrics instrumentations. Specifically, the instrumentation that covers the currently stable OpenTelemetry semantic conventions has been updated to utilize MeterConvention.
The default convention remains the same, although re-implemented using MeterConvention.
An additional implementation for each binder was added that matches the current version (1.37.0) of the OpenTelemetry semantic conventions for JVM metrics.
This allows users to configure Micrometer's JVM metrics instrumentation to use the stable OTel semantic conventions without the need for a `MeterFilter`.
Initial documentation has been added for this, which could be made more thorough in the future.