Use only the main worktree instead of all worktrees in Antora local builds
We set worktrees: true in our antora-playbook.yml which according to the
docs means: "Use all worktrees (the main worktree and all linked
worktrees)". This can break the build if you have a worktree before the
Antora migration. This change sets the default default behavior:
"Use only the main worktree. (default)" since locally that's what you
are interested in anyways.

Closes gh-6823