Fix return type nullability on wrapper methods (#6871)
We were using `@NullUnmarked` which we expected to work to mean the nullability is unknown for the return value. However, without changing the type parameter nullability, it was being considered non-null.

This updates all wrapping methods to use `T extends @Nullable Object` for the type parameter of the return value.

Resolves gh-6869