Remove ThreadLocal usage from LogbackMetrics
Before this change, LogbackMetrics used Counter to keep track of the
number of emitted log events by level. This caused issues since there
are cases where incrementing a Counter can result in emitting
log events. Because of this, the MetricsTurboFilter in LogbackMetrics
had a chance to get into an infinite loop since it was triggered by
emitting log events and it incremented a Counter that also emitted
a log event which triggered MetricsTurboFilter again...

The scenario was the following:
1. Somebody emits a log event (e.g.: LOGGER.info("test");)
2. MetricsTurboFilter gets notified about it and increments a Counter
3. Incrementing the Counter triggers emitting another log event
4. MetricsTurboFilter gets notified about it and increments a Counter
...

To fix this issue, a ThreadLocal was used to ignore some of
the log events (the ones that were emitted by incrementing a Counter).

This change removes the ThreadLocal usage and instead of using Counter,
it keeps track of the number of emitted log events using a LongAdder.
Then it uses this LongAdder in a FunctionCounter. It assumes that
LongAdder does not emit any log events that will end up triggering
MetricsTurboFilter again.