Don't filter log events in LogbackMetricsBenchmark
In LogbackMetricsBenchmark we test LogbackMetrics by emitting log events
on INFO level. In gh-5688 a logback.xml was added that unfortunately set
the lowest log level to WARN and broke LogbackMetricsBenchmark since the
emitted INFO events never reached MetricsTurboFilter in LogbackMetrics
and no metrics were recorded.

Another issue with LogbackMetricsBenchmark is log events going to stdout
even when logback.xml in not present. To solve this we can use Logback's
NOPAppender which will not output anything but lets log events reach
MetricsTurboFilter in LogbackMetrics and record metrics.

See gh-5688