Upgrade to Gradle 9.2.1, add JDK 25 build, upgrade plugins (#6829)
Upgrades to the current latest version of Gradle, which required upgrading a number of Gradle plugin versions and some build configuration. This unlocks building with JDK 25 but the Gradle JVM itself now requires at least 17.

Introduces the `toolchainVersion` property to control what version of the JVM toolchain we use to compile and run tests, separate from the version of the JVM used to run Gradle. Consistently uses JDK 21 to run Gradle because the current Kotlin plugin version does not support running the Gradle JVM with 25.

---------

Signed-off-by: Tommy Ludwig <8924140+shakuzen@users.noreply.github.com>
Co-authored-by: Jonatan Ivanov <jonatan.ivanov@gmail.com>